mascotas y cosas

+++ src/main/java/javaleros/frba/javaleros/repository/MascotaRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MascotaRepository extends JpaRepository<Mascota, Long> {
+public interface MascotaRepository extends JpaRepository<Mascota,Integer> {
 	
 	    
 }
